♻️ refactor: redis src -\> src/common
