+++ server/src/account/account.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Admin } from './account.entity';
 import { AccountRepository } from './account.repository';
 import { winstonModule } from '../common/logger/logger.module';
-import { RedisModule } from '../redis/redis.module';
+import { RedisModule } from '../common/redis/redis.module';
 
 @Module({
   imports: [winstonModule, TypeOrmModule.forFeature([Admin]), RedisModule],
